+++ console.h
@@ -1,4 +1,4 @@
 #include "uart.h"
 
 void show_banner();
-void console_start(size_t buffer_size);
+void console_start(uint64_t buffer_size);
